edge-app/requirements.txt
numpy>=1.21.0
pandas>=1.3.0
requests>=2.26.0
PyYAML>=5.4.1
bleak>=0.14.0  # For Bluetooth LE communication
scipy>=1.7.0   # For signal processing
python-dotenv>=0.19.0 